rename byron core to shelley core
This PR is the first step in a series of four planned PRs to replace
(in the Shelley Ledger code) the module Byron.Spec.Ledger.Core, with a new version
Shelley.Spec.Ledger.Core. The old module will remain unchanged for use in the Byron code.
The new version will be a stripped down version, implementing only the functionality
actualy needed for Shelley, and with more attention paid to efficient impementations.
In all but the last PR, the API of the Relation class, will remain the same.
In this first PR, the Shelley version is an exact copy of the Byron version, except its
module name has been changed to record that it appears in a different place in the source tree.

Future PRs will address the following

step 1) Restrict the export list, and delete the unneeded code.
step 2) Rewrite the needed operations so that they pay attention to efficiency.
step 3) Rework some of the calls to the API, to take advantage of more efficent versions
step 3) Rework the API. Change some inputs from (Foldable f => f a) to (Set a).
        A foldable input forces linear behavior.